fix: added test
